Use REST service with restify

+++ payfast/controllers/pagamentos.js
@@ -45,8 +45,8 @@ module.exports = function( app ) {
 
   app.post( '/pagamentos/pagamento', function( req, res ) {
 
-    req.assert( 'forma_de_pagamento', 'Forma de pagamento é obrigatório' ).notEmpty();
-    req.assert( 'valor', 'Valor é obrigatório e deve ser um decimal' ).notEmpty().isFloat();
+    req.assert( 'pagamento.forma_de_pagamento', 'Forma de pagamento é obrigatório' ).notEmpty();
+    req.assert( 'pagamento.valor', 'Valor é obrigatório e deve ser um decimal' ).notEmpty().isFloat();
 
     var erros = req.validationErrors();
 
@@ -56,7 +56,7 @@ module.exports = function( app ) {
       return;
     }
 
-    var pagamento = req.body;
+    var pagamento = req.body.pagamento;
     console.log( 'Processando uma requisicao de um novo pagamento' );
 
     pagamento.status = 'CRIADO';
@@ -68,30 +68,47 @@ module.exports = function( app ) {
     pagamentoDao.salva( pagamento, function( erro, resultado ){
       if ( erro ) {
         console.log( 'Erro ao inserir no banco' + erro );
-        res.status( 400 ).send( erro );
+        res.status( 500 ).send( erro );
       } else {
         pagamento.id = resultado.insertId;
         console.log( 'pagamento criado' );
 
-        res.location( 'pagamentos/pagamento/' + pagamento.id );
+        if ( pagamento.forma_de_pagamento == 'cartao' ) {
+          var cartao = req.body.cartao;
+          console.log( cartao );
 
-        var response = {
-          dados_do_pagamento: pagamento,
-          links: [
-            {
-              href: 'http://localhost:3000/pagamentos/pagamento/' + pagamento.id,
-              rel: 'confirmar',
-              method: 'PUT'
-            },
-            {
-              href: 'http://localhost:3000/pagamentos/pagamento/' + pagamento.id,
-              rel: 'cancelar',
-              method: 'DELETE'
-            }
-          ]
-        };
+          var clienteCartoes = new app.servicos.clienteCartoes();
 
-        res.status( 201 ).json( response );
+          clienteCartoes.autoriza( cartao, function( exception, request, response, retorno ) {
+            console.log( 'consumindo servico de cartoes' );
+            console.log( retorno );
+
+            res.status( 201 ).json( cartao );
+            return;
+          });
+
+
+        } else {
+          res.location( 'pagamentos/pagamento/' + pagamento.id );
+
+          var response = {
+            dados_do_pagamento: pagamento,
+            links: [
+              {
+                href: 'http://localhost:3000/pagamentos/pagamento/' + pagamento.id,
+                rel: 'confirmar',
+                method: 'PUT'
+              },
+              {
+                href: 'http://localhost:3000/pagamentos/pagamento/' + pagamento.id,
+                rel: 'cancelar',
+                method: 'DELETE'
+              }
+            ]
+          };
+
+          res.status( 201 ).json( response );
+        }
       }
     });
   });
